server/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
httpx>=0.24.0
pillow>=10.0.0
python-multipart>=0.0.6
slowapi>=0.1.9
psnawp-api>=2.0.0
aiofiles>=23.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.0
alembic>=1.12.0
